http://atutor.ca/atutor/mantis/view.php?id=3150

+++ docs/tools/chat/view_transcript.php
@@ -2,7 +2,7 @@
 /************************************************************************/
 /* ATutor																*/
 /************************************************************************/
-/* Copyright (c) 2002-2004 by Greg Gay, Joel Kronenberg & Heidi Hazelton*/
+/* Copyright (c) 2002-2007 by Greg Gay, Joel Kronenberg & Heidi Hazelton*/
 /* Adaptive Technology Resource Centre / University of Toronto			*/
 /* http://atutor.ca														*/
 /*																		*/
@@ -10,13 +10,19 @@
 /* modify it under the terms of the GNU General Public License			*/
 /* as published by the Free Software Foundation.						*/
 /************************************************************************/
-
+// $Id$
 define('AT_INCLUDE_PATH', '../../include/');
 require(AT_INCLUDE_PATH.'vitals.inc.php');
 
-require(AT_INCLUDE_PATH.'header.inc.php');
 
-@readfile(AT_CONTENT_DIR . 'chat/'.$_SESSION['course_id'].'/tran/'.$_GET['t'].'.html');
+$file = AT_CONTENT_DIR . 'chat/'.$_SESSION['course_id'].'/tran/'.$_GET['t'].'.html';
+if (!file_exists($file)) {
+	$msg->addError('FILE_NOT_FOUND');
+	header('Location: index.php');
+	exit;
+}
+require(AT_INCLUDE_PATH.'header.inc.php');
+@readfile($file);
 echo '</table>';
 require(AT_INCLUDE_PATH.'footer.inc.php');
 ?>
